agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "FastAPI Resource Server to validate JWTs from Spring Boot Auth Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "async-timeout>=5.0.1",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.2",
    "fastapi[standard]>=0.120.0",
    "firecrawl>=4.8.0",
    "firecrawl-py>=4.10.4",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=3.0.0",
    "langchain-tavily>=0.2.12",
    "langchain-tools>=0.1.34",
    "langgraph>=1.0.2",
    "lxml>=6.0.2",
    "playwright>=1.56.0",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "tavily-python>=0.7.14",
    "uvicorn[standard]>=0.38.0",
]
